protected routes

+++ src/components/list/list-controller.js
@@ -4,7 +4,7 @@ import TasksModel from "#components/tasks/tasks-model.js";
 
 export async function index (ctx) {
     try {
-        ctx.body = await List.find()
+        ctx.body = await List.find({createBy : ctx.state.user._id})
     } catch (e) {
         ctx.badRequest({message: e.message})
     }
@@ -20,6 +20,7 @@ export async function create(ctx) {
         if (error) {
             throw new Error(error)
         } else {
+            ctx.request.body.createBy = ctx.state.user._id
             List.create(ctx.request.body)
             ctx.status = 204
         }
@@ -30,9 +31,15 @@ export async function create(ctx) {
 }
 
 export async function deleteOne(ctx) {
+
     try {
-        await List.findOneAndRemove(ctx.params.id)
-        ctx.status = 204
+        const list = await List.findOne({_id :ctx.params.id})
+        if(list.createBy.toString() !== ctx.state.user._id.toString()) {
+            ctx.unauthorized()
+        } else {
+            await List.findOneAndRemove(ctx.params.id)
+            ctx.status = 204
+        }
     } catch (e) {
         ctx.badRequest({message : e.message})
     }
@@ -41,9 +48,13 @@ export async function deleteOne(ctx) {
 export async function findOne(ctx) {
     try {
         const list = await List.findById(ctx.params.id).lean()
-        list.tasks = await TasksModel.findByListId(ctx.params.id)
-
-        ctx.body = list
+        if(list.createBy.toString() !== ctx.state.user._id.toString()) {
+            ctx.unauthorized()
+        } else {
+            list.tasks = await TasksModel.findByListId(ctx.params.id)
+            await List.findOneAndRemove(ctx.params.id)
+            ctx.body = list
+        }
     } catch (e) {
         ctx.badRequest({message : e.message})
     }
@@ -51,8 +62,14 @@ export async function findOne(ctx) {
 
 export async function updateOne(ctx) {
     try {
-        await List.findOneAndUpdate(ctx.params.id,  ctx.request.body)
-        ctx.status = 204
+        const list = await List.findById(ctx.params.id).lean()
+        if(list.createBy.toString() !== ctx.state.user._id.toString()) {
+            ctx.unauthorized()
+        } else {
+            ctx.request.body.createBy = ctx.state.user._id
+            await List.findOneAndUpdate(list._id,  ctx.request.body)
+            ctx.status = 204
+        }
     } catch (e) {
         ctx.badRequest({message : e.message})
     }
