+++ src/components/tasks/tasks-controller.js
@@ -3,7 +3,7 @@ import Joi from 'joi'
 
 export async function index (ctx) {
     try {
-        ctx.body = await Tasks.find()
+        ctx.body = await Tasks.find({createBy : ctx.state.user._id})
     } catch (e) {
         ctx.badRequest({message: e.message})
     }
@@ -22,6 +22,7 @@ export async function create(ctx) {
         if (error) {
             throw new Error(error)
         } else {
+            ctx.request.body.createBy = ctx.state.user._id
             Tasks.create(ctx.request.body)
             ctx.status = 204
         }
@@ -33,8 +34,14 @@ export async function create(ctx) {
 
 export async function deleteOne(ctx) {
     try {
-        await Tasks.findOneAndRemove(ctx.params.id)
-        ctx.status = 204
+        const task = await Tasks.findOne({_id :ctx.params.id})
+        if(task.createBy.toString() !== ctx.state.user._id.toString()) {
+            ctx.unauthorized()
+        } else {
+            await Tasks.findOneAndRemove(ctx.params.id)
+            ctx.status = 204
+        }
+
     } catch (e) {
         ctx.badRequest({message : e.message})
     }
@@ -42,7 +49,13 @@ export async function deleteOne(ctx) {
 
 export async function findOne(ctx) {
     try {
-        ctx.body = await Tasks.findById(ctx.params.id)
+        const task = await Tasks.findById(ctx.params.id)
+        if(task.createBy.toString() !== ctx.state.user._id.toString()) {
+            ctx.unauthorized()
+        } else {
+            ctx.body = task
+        }
+
     } catch (e) {
         ctx.badRequest({message : e.message})
     }
@@ -50,7 +63,15 @@ export async function findOne(ctx) {
 
 export async function findAllByListId(ctx) {
     try {
-        ctx.body = await Tasks.findByListId(ctx.params.listId)
+        const taskByList = await Tasks.findByListId(ctx.params.listId)
+
+        taskByList.forEach(task => {
+            if(task.createBy.toString() !== ctx.state.user._id.toString()) {
+                taskByList.splice(1,taskByList.indexOf(task))
+            }
+        })
+        ctx.body = taskByList
+
     } catch (e) {
         ctx.badRequest({message : e.message})
     }
@@ -58,8 +79,13 @@ export async function findAllByListId(ctx) {
 
 export async function updateOne(ctx) {
     try {
-        await Tasks.findOneAndUpdate(ctx.params.id,  ctx.request.body)
-        ctx.status = 204
+        const task = await Tasks.findById(ctx.params.id)
+        if(task.createBy.toString() !== ctx.state.user._id.toString()) {
+            ctx.unauthorized()
+        } else {
+            await Tasks.findOneAndUpdate(task._id,  ctx.request.body)
+            ctx.status = 204
+        }
     } catch (e) {
         ctx.badRequest({message : e.message})
     }
